# doc/api.yaml
openapi: 3.0.0
paths:
  /crash:
    get:
      operationId: AppController_getRoot
      parameters: []
      responses:
        '200':
          description: ''
      tags: &ref_0
        - App
  /unexpected-500-error:
    get:
      operationId: AppController_testUnknownError
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
  /test-logging:
    get:
      operationId: AppController_testLogger
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
  /test-rotation:
    get:
      operationId: AppController_testLogRotation
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
  /user:
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Returns newly created record if request is valid
        '400':
          description: Request body does not contain required field
      summary: Create new user
      tags: &ref_1
        - User
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all users records
      summary: Get all users
      tags: *ref_1
  /user/{id}:
    get:
      operationId: UserController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Returns record with id === userId if it exists
        '400':
          description: UserId is invalid (not uuid)
        '404':
          description: User not found
      summary: Get single user by Id
      tags: *ref_1
    put:
      operationId: UserController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Returns updated record if request is valid
        '400':
          description: UserId is invalid (not uuid)
        '403':
          description: OldPassword is wrong
        '404':
          description: User not found
      summary: Update user's password
      tags: *ref_1
    delete:
      operationId: UserController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Record is found and deleted
        '400':
          description: UserId is invalid (not uuid)
        '404':
          description: User not found
      summary: Delete user
      tags: *ref_1
  /track:
    post:
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: Returns newly created record if request is valid
        '400':
          description: Request body does not contain required field
      summary: Create new track
      tags: &ref_2
        - Track
    get:
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all tracks records
      summary: Get all tracks
      tags: *ref_2
  /track/{id}:
    get:
      operationId: TrackController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Returns record with id === trackId if it exists
        '400':
          description: TrackId is invalid (not uuid)
        '404':
          description: Track not found
      summary: Get single track by Id
      tags: *ref_2
    put:
      operationId: TrackController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: Returns updated record if request is valid
        '400':
          description: TrackId is invalid (not uuid)
        '404':
          description: Track not found
      summary: Update track info
      tags: *ref_2
    delete:
      operationId: TrackController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Record is found and deleted
        '400':
          description: TrackId is invalid (not uuid)
        '404':
          description: Track not found
      summary: Delete track
      tags: *ref_2
  /artist:
    post:
      operationId: ArtistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Returns newly created record if request is valid
        '400':
          description: Request body does not contain required field
      summary: Create new artist
      tags: &ref_3
        - Artist
    get:
      operationId: ArtistController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all artists records
      summary: Get all artists
      tags: *ref_3
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Returns record with id === artistId if it exists
        '400':
          description: ArtistId is invalid (not uuid)
        '404':
          description: Artist not found
      summary: Get single artist by Id
      tags: *ref_3
    put:
      operationId: ArtistController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Returns updated record if request is valid
        '400':
          description: ArtistId is invalid (not uuid)
        '404':
          description: Artist not found
      summary: Update artist info
      tags: *ref_3
    delete:
      operationId: ArtistController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Record is found and deleted
        '400':
          description: ArtistId is invalid (not uuid)
        '404':
          description: Artist not found
      summary: Delete artist
      tags: *ref_3
  /album:
    post:
      operationId: AlbumController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Returns newly created record if request is valid
        '400':
          description: Request body does not contain required field
      summary: Create new album
      tags: &ref_4
        - Album
    get:
      operationId: AlbumController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all albums records
      summary: Get all albums
      tags: *ref_4
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Returns record with id === albumId if it exists
        '400':
          description: AlbumId is invalid (not uuid)
        '404':
          description: Album not found
      summary: Get single album by Id
      tags: *ref_4
    put:
      operationId: AlbumController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Returns updated record if request is valid
        '400':
          description: AlbumId is invalid (not uuid)
        '404':
          description: Album not found
      summary: Update album info
      tags: *ref_4
    delete:
      operationId: AlbumController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Record is found and deleted
        '400':
          description: AlbumId is invalid (not uuid)
        '404':
          description: Album not found
      summary: Delete album
      tags: *ref_4
  /favs:
    get:
      operationId: FavsController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all favorite records (not their ids), split by entity type
      summary: Get all favorites
      tags: &ref_5
        - Favs
  /favs/track/{id}:
    post:
      operationId: FavsController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Track exists and added to favorites
        '400':
          description: TrackId is invalid (not uuid)
        '422':
          description: Track with id === trackId does not exist
      summary: Add track to the favorites
      tags: *ref_5
    delete:
      operationId: FavsController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Track was in favorites and now it is deleted
        '400':
          description: TrackId is invalid (not uuid)
        '404':
          description: Track is not favorite
      summary: Delete track from favorites
      tags: *ref_5
  /favs/artist/{id}:
    post:
      operationId: FavsController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Artist exists and added to favorites
        '400':
          description: ArtistId is invalid (not uuid)
        '422':
          description: Artist with id === albumId does not exist
      summary: Add artist to the favorites
      tags: *ref_5
    delete:
      operationId: FavsController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Artist was in favorites and now it is deleted
        '400':
          description: ArtistId is invalid (not uuid)
        '404':
          description: Artist is not favorite
      summary: Delete artist from favorites
      tags: *ref_5
  /favs/album/{id}:
    post:
      operationId: FavsController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Album exists and added to favorites
        '400':
          description: AlbumId is invalid (not uuid)
        '422':
          description: Album with id === albumId does not exist
      summary: Add album to the favorites
      tags: *ref_5
    delete:
      operationId: FavsController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Album was in favorites and now it is deleted
        '400':
          description: AlbumId is invalid (not uuid)
        '404':
          description: Album is not favorite
      summary: Delete album from favorites
      tags: *ref_5
  /auth/signup:
    post:
      operationId: AuthController_signup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Record created
        '400':
          description: Dto is invalid (no login or password, or they are not a strings
      summary: Register new user
      tags: &ref_6
        - Auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserDto'
      responses:
        '200':
          description: Return tokens
        '400':
          description: Dto is invalid (no login or password, or they are not a strings
        '403':
          description: >-
            Authentication failed (no user with such login, password does not
            match actual one, etc.)
      summary: Get access and refresh token
      tags: *ref_6
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Tokens refreshed successfully
        '401':
          description: No refresh token provided
        '403':
          description: Invalid or expired refresh token
      summary: >-
        Send refresh token in body as { refreshToken } to get new pair of Access
        token and Refresh token
      tags: *ref_6
info:
  title: Home Library Service
  description: Home music library service API
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: Login
          description: User login
        password:
          type: string
          example: Password123
          description: User password
      required:
        - login
        - password
    UpdateUserDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: OldPassword123
          description: User old password
        newPassword:
          type: string
          example: NewPassword123
          description: User new password
      required:
        - oldPassword
        - newPassword
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: My track
          description: Track name
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: Artist uuid
          nullable: true
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: Album uuid
          nullable: true
        duration:
          type: number
          example: '60'
          description: Duration
      required:
        - name
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: My track
          description: Track name
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: Artist uuid
          nullable: true
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: Album uuid
          nullable: true
        duration:
          type: number
          example: '60'
          description: Duration
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: John Doe
          description: Artist name
        grammy:
          type: boolean
          example: true
          description: Flag if artist has grammy
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: John Doe
          description: Artist name
        grammy:
          type: boolean
          example: true
          description: Flag if artist has grammy
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: My album
          description: Album name
        year:
          type: number
          example: 2025
          description: Release year
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: Artist uuid
          nullable: true
      required:
        - name
        - year
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: My album
          description: Album name
        year:
          type: number
          example: 2025
          description: Release year
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: Artist uuid
          nullable: true
    LoginUserDto:
      type: object
      properties:
        login:
          type: string
          example: Login
          description: User login
        password:
          type: string
          example: Password123
          description: User password
      required:
        - login
        - password
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: refresh_token_string
          description: Refresh token